sincronizacion y persistencia de datos
