+++ amplify/backend/api/testappwom1/schema.graphql
@@ -1,7 +1,7 @@
-type TareasPendientes @model @auth(rules: [{allow: public}]) {
+type Todo @model @auth(rules: [{allow: public}]) {
   id: ID!
-  name: String
-  descripcion: String
-  is_complete: Boolean
+  name: String!
+  description: String
+  isComplete: Boolean!
 }
