+++ amplify/backend/api/testappwom1/transform.conf.json
@@ -1,10 +1,4 @@
 {
     "Version": 5,
-    "ElasticsearchWarning": true,
-    "ResolverConfig": {
-        "project": {
-            "ConflictHandler": "AUTOMERGE",
-            "ConflictDetection": "VERSION"
-        }
-    }
+    "ElasticsearchWarning": true
 }
