+++ amplify/backend/backend-config.json
@@ -20,5 +20,13 @@
         }
       }
     }
+  },
+  "auth": {
+    "testappwom14db08f48": {
+      "service": "Cognito",
+      "providerPlugin": "awscloudformation",
+      "dependsOn": [],
+      "customAuth": false
+    }
   }
 }
