+++ pubspec.yaml
@@ -32,6 +32,8 @@ dependencies:
   firebase_core: ^1.3.0
   amplify_flutter: ^0.2.0
   amplify_datastore: ^0.2.0
+  amplify_api: ^0.2.0
+  amplify_auth_cognito: ^0.2.0
 
 dev_dependencies:
   flutter_test:
